[naming] rename filegen to shellgen; and move devcontainer_util.go to /impl/generate
