+++ internal/impl/generate/tmpl/devcontainerDockerfile.tmpl
