+++ internal/impl/generate/tmpl/envrcContent.tmpl
