+++ internal/shellgen/doc.go
@@ -1,0 +1,6 @@
+package shellgen
+
+// shellgen package is responsible for printing files in the .devbox/gen directory that
+// are needed to create the "devbox shell environment".
+//
+// This is flake.nix, and the init-hooks and scripts from the devbox.json.
