+++ internal/shellgen/flake_input.go
@@ -1,4 +1,4 @@
-package filegen
+package shellgen
 
 import (
 	"context"
