+++ internal/shellgen/flake_plan.go
@@ -1,4 +1,4 @@
-package filegen
+package shellgen
 
 import (
 	"context"
