+++ internal/shellgen/generate_test.go
@@ -1,7 +1,7 @@
 // Copyright 2023 Jetpack Technologies Inc and contributors. All rights reserved.
 // Use of this source code is governed by the license in the LICENSE file.
 
-package filegen
+package shellgen
 
 import (
 	"flag"
