+++ internal/shellgen/testdata/flake-empty.nix.golden
