+++ internal/shellgen/testdata/flake-with-global.nix.golden
