+++ internal/shellgen/testdata/flake.nix.golden
