+++ internal/shellgen/tmpl/.gitignore.tmpl
