+++ internal/shellgen/tmpl/flake.nix.tmpl
