+++ internal/shellgen/tmpl/shell.nix.tmpl
